ANPR/requirements.txt
easyocr==1.7.0
Flask==2.3.2
matplotlib==3.7.2
numpy==1.25.2
opencv_contrib_python==4.8.0.74
opencv_python_headless==4.8.0.74
Pillow==9.5.0
torch==2.0.1+cu118
ultralytics==8.0.148
